Distance, Amount and Duration calculation added in End Ride API

+++ src/main/java/com/ridesharing/drivermanagementservice/models/Ride.java
@@ -24,7 +24,7 @@ public class Ride extends BaseDriverModel {
     private String riderLastName;
     private String riderPhone;
     private String status;
-    private Float distance;
+    private Double distance;
     private Float amount;
     private Integer duration;
